Always read gRPC token file
Currently the controller only reads the file containing the token used
to authenticate to the fulfillment service once, during startup. This
means that it will not automatically get new tokens when the expire and
the file is overwritten. To avoid that this patch changes the code so
that it always read the token from the file before using it.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>