Trigger the webhook on every reconcile
Previously, we were only calling the AAP webhook if the HostedCluster
resource did not exist. This was an attempt to deal with concurrency issues
we were experiencing in AAP, but it prevents us from being able to update
the cluster after it has been created (specifically, it means we are
unable to update the number of nodes associated with the cluster).

With this change, we trigger the webhook on every reconcile, and rely on
the concurrency controls we have implemented in the Ansible roles.