Fix webhook rate limiting
The webhook code used only the resource name as the cache key. This caused
some requests to get rate limited erroneously -- for example, when
attempting to delete a cluster, the initial delete request would get
dropped if there had been a recent create request.

This commit updates the webhook client to use both the url and resource
name in the cache key.

Closes: #74