Always trigger delete webhook when deleting
Previously, we would only call the delete webhook if (a) the working name space
exists and (b) the hostedcluster exists. This prevented us from re-trying
the delete job to clean up openstack resources (networks, floating ips,
etc).

With this change, we always call the delete webhook as part of handling
delete events.