ci: Add release workflow with automatic versioning
Add a workflow that uses conventional commits [1] to automatically
increment the version, and generate a new release on GitHub when there is a
new version number.

In short:

- Commits prefixed with `fix:` will increment the patch version number
  (`v0.1.0 -> v0.1.1`).

- Commits prefixed with `feat:` will increment the minor version number
  (`v0.1.0 -> v0.2.0`)

- Commits that introduce a breaking change (typically signified with a `!`
  after the commit type, like `feat!:`) will increment the major version
  number.

  You can also introduce a breaking change by including a `BREAKING
  CHANGE:` footer in your commit message.

We build a new container image on every push. For releases, we tag the
image with the version number (and the minor version, and the major
version, so `v0.1.1` will result in an image with tags `v0`, `v0.1`, and
`v0.1.1`) and update the `latest` tag to point to the new image.

Pushes that do not include any conventional commit tags will not generate a
new release.

[1]: https://www.conventionalcommits.org/en/v1.0.0/

x-branch: feature/release-workflow