Prevent e2e and integration tests from running by default
We have a bunch of integration tests that spin up an etcd and api-server,
and it's good that we have them, but they are unsuitable for running as
part of a pre-commit hook (or in other situations in which a longer running
test can be disruptive).

This commit marks all integration tests with the `integration` build tag.
That makes it possible to run `go test ./...` and have it complete
successfully.

If you want to run the integration tests, you can run:

    make test-integration

Or if you have the binary requirements already in place, you can run:

    go test -tags=integration ./...

The e2e-tests are protected by the `e2e` tag. You can run them via:

    make test-e2e

(This is unchanged, although we have modified the Makefile to use build
tags rather than `grep` to target specific tests.)

x-branch: feature/disable-e2e